LibPDF: Remove the assumption that everything has to be an object
For example, PDF::IndirectObject only wrapped another object, not a
value. This means that the following object would crash:

    1 0 obj
    12345
    endobj

This is just a plain number, which is just a Value, not an object.
There are other locations where this was a problem as well, such as
the document's object cache.

Additionally, Object#resolve{,to} has been moved to Document, and takes
a Value. This allows Values to be resolved as well (and to the
primitive values), which increases code quality.