LibWeb: Don't skip returning initial value when collecting animations
The original fix did not call maybe_null_property() since that method
was changed to return the result of animation calculation, if one
exists, before returning the underlying value. However, this broke
some keyframes without "from" and "to" frames. The fix here is to add
a method that always returns the underlying style value and ignores
animated values entirely.