ClangPlugins: Suppress capture warnings with annotation on lambda
Instead of annotating variable declarations with
IGNORE_USE_IN_ESCAPING_LAMBDA, lambdas are now annotated with
DOES_NOT_OUTLIVE_CAPTURES, which suppresses warnings for the entire
capture list instead of just a single variable. This makes more sense,
as if one capture is not outlived by the lambda, it goes to follow that
none of them are.

Note that the annotation placement is a little strange though. There are
two places that annotations can be placed:

    [] [[first_annotation]] (...) -> ... [[second_annotation]] {};

Above, [[first_annotation]] is applied to the operator() of the
generated lambda struct. As it is a declaration, it must be a
clang::annotate attribute. [[second_annotation]] is applied to the
return type of operator(), and thus must be a clang::annotate_type
attribute. As we want to annotate the lambda itself, the
DOES_NOT_OUTLIVE_CAPTURES annotation must be placed in the first
position.